using redis for sessions

+++ services/gateway/deps/jwt.py
@@ -7,7 +7,7 @@ KEY = "MANTAP"
 
 class JwtHelper:
     def encode(self, other_data: dict) -> str:
-        return jwt.encode({"exp": time.time() + (1 * 3600), **other_data}, KEY)
+        return jwt.encode({"exp": int(time.time() + (1 * 3600)), **other_data}, KEY)
 
     def decode(self, token: str) -> dict:
         return jwt.decode(token.encode(), KEY, 'HS256')
